+++ tests/Unit/Private/Get-DaikinBasicInfo.Tests.ps1
@@ -1,0 +1,55 @@
+
+BeforeDiscovery {
+    $ProjectPath = "$PSScriptRoot\..\..\.." | Convert-Path
+    $ProjectName = (Get-ChildItem $ProjectPath\*\*.psd1 | Where-Object {
+        ($_.Directory.Name -eq 'source') -and
+            $(try
+                {
+                    Test-ModuleManifest $_.FullName -ErrorAction Stop
+                }
+                catch
+                {
+                    $false
+                }) }
+    ).BaseName
+
+    Import-Module $ProjectName
+}
+
+InModuleScope $ProjectName {
+    Describe -Name 'Get-DaikinBasicInfo.ps1' -Fixture {
+        Context -Name 'When retreival succeeds' {
+            BeforeAll {
+                Mock Invoke-RestMethod -MockWith {}
+                function Convert-DaikinResponse
+                {
+                }
+                Mock Convert-DaikinResponse -MockWith { return [ordered]@{} }
+            }
+            It -Name 'Should not throw' {
+                { Get-DaikinBasicInfo -Hostname 'daikin.network.com' } | Should -Not -Throw
+            }
+        }
+        Context -Name 'When Invoke-RestMethod fails' {
+            BeforeAll {
+                Mock Invoke-RestMethod -MockWith { throw }
+            }
+            It -Name 'Should throw' {
+                { Get-DaikinBasicInfo -Hostname 'daikin.network.com' } | Should -Throw
+            }
+        }
+        Context -Name 'When Convert-DaikinResponse fails' {
+            BeforeAll {
+                Mock Invoke-RestMethod -MockWith {}
+                function Convert-DaikinResponse
+                {
+                }
+                Mock Convert-DaikinResponse -MockWith { throw }
+            }
+            It -Name 'Should throw' {
+                { Get-DaikinBasicInfo -Hostname 'daikin.network.com' } | Should -Throw
+            }
+        }
+    }
+
+}