+++ tests/Unit/Private/Invoke-GarbageCollect.tests.ps1
@@ -1,4 +1,4 @@
-BeforeAll {
+BeforeDiscovery {
     $ProjectPath = "$PSScriptRoot\..\..\.." | Convert-Path
     $ProjectName = (Get-ChildItem $ProjectPath\*\*.psd1 | Where-Object {
         ($_.Directory.Name -eq 'source') -and
@@ -19,7 +19,7 @@ InModuleScope $ProjectName {
     Describe Invoke-GarbageCollect {
         Context 'Default' {
             It 'Should not throw' {
-                {Invoke-GarbageCollect} | Should -not -throw
+                { Invoke-GarbageCollect } | Should -Not -Throw
             }
         }
     }